requirements.txt
PyQt6>=6.5.0
aiogram>=3.0.0
google-generativeai>=0.3.0
aiohttp>=3.9.0
Pillow>=10.0.0
pyautogui>=0.9.54